port all existing auth logic to tRPC
