--- backend/src/auth/schemas.ts
@@ -1,13 +1,0 @@
-import { z } from 'zod';
-
-export const loginSchema = z.object({
-  email: z.string().email(),
-  password: z.string(),
-});
-
-export const registerSchema = z.object({
-  firstName: z.string().optional(),
-  lastName: z.string().optional(),
-  email: z.string().email(),
-  password: z.string(),
-});
